Rough draft of time-filtering gesture
The time-filtering gesture now works, but with a serious bug that repros
as follows: first, build the graph

(view-channel c1)

Next, remove the edge to produce

view-channel c1

Finally, drag a filter selection on the view (which is still showing,
because we haven't entered a valid state since it was drawn.) Since
there are no incoming edges, the filter is not added.

A correct implementation probably involves moving _addFilter() to
ViewGraph, passing a complete fist program through the filteradded
event, and reconstructing the entire ViewGraph from scratch. The visual
jitter shouldn't be too bad, and this will also make it cleaner to
determine block sizes from filter text.