Fixing empty channel bug
d3.min([]) === undefined, which means that viewing an empty channel
bounced undefined and NaN values all around the view code. This
invariably leads to TypeError, which is fixed in this diff by enforcing
numeric bounds.

A better solution is to infer the correct bounds based on any filters
applied in the S-expression. That will have to wait.